Banner text tweak, Other small text changes
Initial banner text change was in previous commit (I failed to note that in the commit notes). This fixes a type / missed word in that one. 
Updated text in "You can't see or taste it" section
In the "We Partner" section: changed "doctors" --> "pediatricians". Bolded that full sentence rather than just a section.